Make the regex more relaxed in the test
Also replace 'error' with 'warn'

We are passing the string to is-ignored, where it is passed to grep, so
we have to be careful with any special characters that can get lost or
lose its escaping on the way, so in the test it's good to just use a
loose regex and use dots, for example.

Our test log is very limited, we know what is in there.