Support additional arguments in logwarn
As per the documentation there are options as `-T` which have to come after
the filename. Here the script is modified to facilitate this extra options.
Maybe this will need to be adjust also, in a way that applies to specific
patterns in the future.As for not it applies to all subsequent patterns if it
is set.

Signed-off-by: Ioannis Bonatakis <ybonatakis@suse.com>